Moved agents models to dedicated file.
